+++ unlock-js/CHANGELOG.md
@@ -3,6 +3,7 @@
 # Next version
 - Code cleaned witth automated linting fixes
 - Adjusted lint config to ignore violations that could not be fixed auttomatically
+- Added setKeyMetadata and getKeyMetadata methods to WalletService
 
 # 0.7.0